Added template to show collection view
The collection view is different from the view for other taxonomy terms.
Others are grouped by collections but Mike did not want the collection view
to be grouped by collection because this has the effect of showing collections
other than the one clicked on, due to some terms being in more than one.

I looked at other ways of implementing different views for taxonomy
terms from different vocabs. I tried changing the url alias to include the vocab
name. I then tried creating a view with a matching path in the hope it would
override it but it did not. I tried creating a block of the view but the block
placement config is not fine-grained enough given that I needed to show the
default taxonomy view on all taxonomy pages except collection & topic map.
So we are stuck with this.

This being in the topic map module could be a problem if other people
want to use the module since they won't have collections or a collection
view. However it might not throw any errors since the template suggestion
will never be matched.